+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Allow `gemini:` and `monero:` links in post content.
 - Update `monerod` and `monero-wallet-rpc` containers.
+- Check chain ID consistency when reading/writing payment options.
 
 ## [2.0.0] - 2023-10-02